# docs/openapi/reconcile-v1.yaml
openapi: 3.1.0
info:
  title: Reconcile - payment & settlement reconciliation
  description: |
    Money movement, provider settlement ingestion, and reconciliation.

    Every mutating endpoint requires an `Idempotency-Key` and replays a
    stored response for 24 hours. Refusals are RFC 9457 problem
    documents; the failure taxonomy is in `docs/spec/07-failure-matrix.md`
    and the error table in `docs/spec/06-api-contract.md`.

    Two routes are deliberately unauthenticated: `GET /api/v1/health`,
    which a load balancer must be able to reach, and
    `POST /api/v1/provider/webhooks`, which authenticates by HMAC
    signature rather than by token.
  license:
    name: Proprietary
  version: v1
servers:
- url: /
  description: This application
security:
- bearerAuth: []
paths:
...
  /api/v1/reconciliation/cases/{id}/resolve:
    post:
      tags:
      - reconciliation-controller
      operationId: resolve
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveRequest"
        required: true
      responses:
        "200":
          description: OK
...
  /api/v1/payments:
    get:
      tags:
      - payment-controller
      operationId: list_1
      parameters:
      - name: state
        in: query
        required: false
        schema:
          type: string
      - name: merchantReference
        in: query
        required: false
        schema:
          type: string
      - name: providerTransactionId
        in: query
        required: false
        schema:
          type: string
      - name: cursor
        in: query
        required: false
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponsePaymentResponse"
    post:
      tags:
      - payment-controller
      operationId: create
      parameters:
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /api/v1/payments/{id}/refund:
    post:
      tags:
      - payment-controller
      operationId: refund
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /api/v1/payments/{id}/fail:
    post:
      tags:
      - payment-controller
      operationId: fail
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FailRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /api/v1/payments/{id}/capture:
    post:
      tags:
      - payment-controller
      operationId: capture
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaptureRequest"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /api/v1/payments/{id}/authorize:
    post:
      tags:
      - payment-controller
      operationId: authorize
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
...
  /api/v1/payments/{id}:
    get:
      tags:
      - payment-controller
      operationId: get_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
  /api/v1/payments/{id}/timeline:
    get:
      tags:
      - payment-controller
      operationId: timeline
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimelineResponse"
...
components:
  schemas:
...
    ResolveRequest:
      type: object
      properties:
        resolutionNote:
          type: string
          maxLength: 2000
          minLength: 20
        adjustmentLedgerTransactionId:
          type: string
      required:
      - resolutionNote
...
    CreatePaymentRequest:
      type: object
      properties:
        merchantReference:
          type: string
          minLength: 1
          pattern: "^[A-Za-z0-9_-]{1,64}$"
        description:
          type: string
          maxLength: 280
          minLength: 0
        amount:
          $ref: "#/components/schemas/MoneyDto"
        provider:
          type: string
        providerTransactionId:
          type: string
      required:
      - amount
      - merchantReference
    MoneyDto:
      type: object
      properties:
        amountMinor:
          type: integer
          format: int64
        currency:
          type: string
    FailRequest:
      type: object
      properties:
        failureCode:
          type: string
          maxLength: 32
          minLength: 0
        reason:
          type: string
          maxLength: 280
          minLength: 0
      required:
      - failureCode
    CaptureRequest:
      type: object
      properties:
        expectedAmount:
          $ref: "#/components/schemas/MoneyDto"
...
    PageResponsePaymentResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PaymentResponse"
        nextCursor:
          type: string
    PaymentResponse:
      type: object
      properties:
        id:
          type: string
        merchantReference:
          type: string
        description:
          type: string
        amount:
          $ref: "#/components/schemas/MoneyDto"
        platformFee:
          $ref: "#/components/schemas/MoneyDto"
        netAmount:
          $ref: "#/components/schemas/MoneyDto"
        state:
          type: string
        failureCode:
          type: string
        failureReason:
          type: string
        provider:
          type: string
        providerTransactionId:
          type: string
        capturedAt:
          type: string
          format: date-time
        settledAt:
          type: string
          format: date-time
        paidOutAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
    TimelineResponse:
      type: object
      properties:
        occurredAt:
          type: string
          format: date-time
        kind:
          type: string
        reference:
          type: string
        summary:
          type: string
        requestId:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      description: |
        A static operator or admin bearer token. Admin implies
        operator. Not a production identity design; see the
        authentication section of the README for what a real
        deployment puts in front of this instead.
      scheme: bearer